Moved filter by member in a separate file
